Move some ssse3 intrinsics from i586 to i686
